pyproject.toml
[project]
name = "confluence-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "atlassian-python-api>=4.0.7",
    "beautifulsoup4>=4.14.3",
    "faiss-cpu>=1.13.2",
    "fastapi>=0.135.3",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.2",
    "langchain-ollama>=1.1.0",
    "langchain-text-splitters>=1.1.1",
    "pydantic-settings>=2.13.1",
    "requests>=2.33.1",
    "streamlit>=1.56.0",
    "uvicorn>=0.44.0",
]
